+++ com.unity.visualeffectgraph/Editor/Models/Contexts/VFXSubgraphContext.cs
@@ -105,7 +105,6 @@ namespace UnityEditor.VFX
 
         public override VFXExpressionMapper GetExpressionMapper(VFXDeviceTarget target)
         {
-            PatchInputExpressions();
             return null;
         }
 
@@ -195,10 +194,10 @@ namespace UnityEditor.VFX
                     }
                 }
                 else if( child.copy is VFXSubgraphBlock subgraphBlock)
-                {
-                    subgraphBlock.RecreateCopy();
+                    {
+                        subgraphBlock.RecreateCopy();
+                    }
                 }
-            }
 
             List<string> newInputFlowNames = new List<string>();
 
@@ -270,7 +269,6 @@ namespace UnityEditor.VFX
                             LinkFrom(link.context,link.slotIndex,i);
                 }
             }
-            PatchInputExpressions();
             SyncSlots(VFXSlot.Direction.kInput,true);
         }
 
@@ -336,7 +334,7 @@ namespace UnityEditor.VFX
 
         Dictionary<VFXSlot, VFXSlot> m_OriginalToCopy = new Dictionary<VFXSlot, VFXSlot>();
 
-        void PatchInputExpressions()
+        public void PatchInputExpressions()
         {
             if (m_SubChildren == null) return;